+++ src/filters/FilterBackend.ts
@@ -6,6 +6,8 @@ import { webGLProbe } from './WebGLProbe';
 
 export type FilterBackend = WebGLFilterBackend | Canvas2dFilterBackend;
 
+let filterBackend: FilterBackend;
+
 export function initFilterBackend(): FilterBackend {
   webGLProbe.queryWebGL();
   if (config.enableGLFiltering && webGLProbe.isSupported(config.textureSize)) {
@@ -16,15 +18,18 @@ export function initFilterBackend(): FilterBackend {
 }
 
 /**
- * @todo refactor to a module w/o assigning to fabric
+ *
+ * @param [strict] pass `true` to create the backend if it wasn't created yet (default behavior),
+ * pass `false` to get the backend ref without mutating it
  */
-export function getFilterBackend(): FilterBackend {
-  if (!fabric.filterBackend) {
-    fabric.filterBackend = initFilterBackend();
+export function getFilterBackend(strict = true): FilterBackend {
+  if (!filterBackend && strict) {
+    filterBackend = initFilterBackend();
   }
-  return fabric.filterBackend;
+  return filterBackend;
 }
 
+fabric.getFilterBackend = getFilterBackend;
 fabric.Canvas2dFilterBackend = Canvas2dFilterBackend;
 fabric.WebglFilterBackend = WebGLFilterBackend;
 fabric.initFilterBackend = initFilterBackend;